🍱 Add: 회원 탈퇴 사유 추가 (#50)
- 회원 탈퇴 사유 DTO
- 회원 탈퇴 사유 이넘 클래스 추가
- 회원 탈퇴 유저 엔티티, 레포지토리 추가
- 기타 사유면 reason를 쓰도록 예외처리

+++ src/main/kotlin/com/eatsfinder/domain/user/service/UserServiceImpl.kt
@@ -5,7 +5,10 @@ import com.eatsfinder.domain.email.service.EmailUtils
 import com.eatsfinder.domain.post.repository.PostRepository
 import com.eatsfinder.domain.user.dto.user.*
 import com.eatsfinder.domain.user.dto.user.active.MyActiveResponse
+import com.eatsfinder.domain.user.model.DeleteUserData
 import com.eatsfinder.domain.user.model.SocialType
+import com.eatsfinder.domain.user.model.DeleteUserReason
+import com.eatsfinder.domain.user.repository.DeleteUserDataRepository
 import com.eatsfinder.domain.user.repository.UserLogRepository
 import com.eatsfinder.domain.user.repository.UserRepository
 import com.eatsfinder.global.aws.AwsS3Service
@@ -13,10 +16,7 @@ import com.eatsfinder.global.exception.ModelNotFoundException
 import com.eatsfinder.global.exception.email.ExpiredCodeException
 import com.eatsfinder.global.exception.email.NotCheckCompleteException
 import com.eatsfinder.global.exception.email.OneTimeMoreWriteException
-import com.eatsfinder.global.exception.profile.AlreadyDefaultProfileImageException
-import com.eatsfinder.global.exception.profile.ImmutableUserException
-import com.eatsfinder.global.exception.profile.MyProfileException
-import com.eatsfinder.global.exception.profile.WrongPasswordException
+import com.eatsfinder.global.exception.profile.*
 import com.eatsfinder.global.security.jwt.UserPrincipal
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.crypto.password.PasswordEncoder
@@ -31,6 +31,7 @@ class UserServiceImpl(
     private val passwordEncoder: PasswordEncoder,
     private val postRepository: PostRepository,
     private val userLogRepository: UserLogRepository,
+    private val deleteUserDataRepository: DeleteUserDataRepository,
     private val emailUtils: EmailUtils,
     private val awsService: AwsS3Service
 ) : UserService {
@@ -123,22 +124,32 @@ class UserServiceImpl(
         }
     }
 
-    override fun deleteProfile(myProfileId: Long, email: String, code: String) {
+    override fun deleteProfile(myProfileId: Long, req: DeleteReasonRequest, reasonType: DeleteUserReason) {
         val profile = userRepository.findByIdAndDeletedAt(myProfileId, null) ?: throw ModelNotFoundException(
             "user",
             "이 프로필은(id: ${myProfileId})은 존재하지 않습니다."
         )
+        if (reasonType.name == DeleteUserReason.OTHERS.name && req.reason == null){
+            throw EnterAddInfoException("기타 사유를 입력해주세요")
+        }
 
-        val checkCode = emailRepository.findByCode(code)
+        val checkCode = emailRepository.findByCode(req.code)
         when {
-            checkCode == null || !(checkCode.code == code && profile.email == checkCode.email && profile.email == email) -> throw OneTimeMoreWriteException("다시 한번 입력해주세요")
+            checkCode == null || !(checkCode.code == req.code && profile.email == checkCode.email && profile.email == req.email) -> throw OneTimeMoreWriteException("다시 한번 입력해주세요")
             checkCode.expiredAt.isBefore(LocalDateTime.now()) -> throw ExpiredCodeException("인증번호가 만료되었습니다.")
             !checkCode.complete-> throw NotCheckCompleteException(
                 "인증확인이 되지 않았습니다."
             )
             else -> {
                 userRepository.delete(profile)
-                emailUtils.guideEmail(email)
+                emailUtils.guideEmail(req.email)
+                deleteUserDataRepository.save(
+                    DeleteUserData(
+                        userId = profile,
+                        reason = req.reason,
+                        reasonType = reasonType
+                    )
+                )
             }
         }
     }
